added vehicle category config and refactored UI with services

+++ src/main/java/org/example/services/UserService.java
@@ -1,7 +1,7 @@
 package org.example.services;
 
+import org.example.models.Role;
 import org.example.models.User;
-import org.example.models.Vehicle;
 import org.example.repositories.UserRepository;
 
 import java.util.List;
@@ -10,20 +10,32 @@ import java.util.Optional;
 public class UserService {
     private final UserRepository userRepository;
 
-    public UserService(UserRepository userRepository) {
+    private final RentalService rentalService;
+
+    public UserService(UserRepository userRepository, RentalService rentalService) {
         this.userRepository = userRepository;
+        this.rentalService = rentalService;
     }
 
-    public List<User> getUsers(){
+    public List<User> findAllUsers(){
         return this.userRepository.findAll();
     }
 
-    public User deleteUser(String userId){
+    public User findById(String userId){
+        return this.userRepository.findById(userId).get();
+    }
+
+    public void deleteUser(String userId, String adminId){
+        if(!findById(adminId).getRole().equals(Role.ADMIN))
+            throw new IllegalArgumentException("Deleting user must be admin");
+
+        if(rentalService.userHasActiveRental(userId))
+            throw new IllegalArgumentException("This user is renting");
+
         Optional<User> opt = this.userRepository.findById(userId);
         if(opt.isEmpty())
             throw new IllegalArgumentException("No user with such ID");
 
         this.userRepository.deleteById(userId);
-        return opt.get();
     }
 }
