+++ src/main/java/org/example/services/VehicleService.java
@@ -1,8 +1,10 @@
 package org.example.services;
 
 import org.example.models.Vehicle;
+import org.example.repositories.RentalRepository;
 import org.example.repositories.VehicleRepository;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -10,29 +12,52 @@ import java.util.Optional;
 public class VehicleService {
     private final VehicleRepository vehicleRepository;
 
-    public VehicleService(VehicleRepository vehicleRepository) {
+    private final RentalRepository rentalRepository;
+
+    private final VehicleValidator vehicleValidator;
+
+    public VehicleService(VehicleRepository vehicleRepository, RentalRepository rentalRepository, VehicleValidator vehicleValidator) {
         this.vehicleRepository = vehicleRepository;
+        this.rentalRepository = rentalRepository;
+        this.vehicleValidator = vehicleValidator;
     }
 
-    public List<Vehicle> getVehicles(){
+    public List<Vehicle> findAllVehicles(){
         return this.vehicleRepository.findAll();
     }
 
-    public boolean checkVehicle(String vehicleId){
-        return this.vehicleRepository.findById(vehicleId).isPresent();
+    public List<Vehicle> findAvailableVehicles(){
+        List<Vehicle> all = this.vehicleRepository.findAll();
+        List<Vehicle> available = new ArrayList<>();
+        for(Vehicle v : all){
+            if(!isVehicleRented(v.getId())){
+                available.add(v);
+            }
+        }
+        return available;
     }
 
-    public Vehicle addVehicle(String category, String brand, String model, int year, String plate, double price, Map<String, Object> attributes){
-        Vehicle vehicle = new Vehicle(null, category, brand, model, year, plate, price, attributes);
+    public Vehicle findById(String vehicleId){
+        return this.vehicleRepository.findById(vehicleId).get();
+    }
+
+    public boolean isVehicleRented(String vehicleId){
+        return this.rentalRepository.findByVehicleIdAndReturnDateIsNull(vehicleId).isPresent();
+    }
+
+    public Vehicle addVehicle(Vehicle vehicle){
+        this.vehicleValidator.validate(vehicle);
         return this.vehicleRepository.save(vehicle);
     }
 
-    public Vehicle deleteVehicle(String vehicleId) {
+    public void removeVehicle(String vehicleId) {
+        if(isVehicleRented(vehicleId))
+            throw new IllegalArgumentException("This vehicle is rented");
+
         Optional<Vehicle> opt = this.vehicleRepository.findById(vehicleId);
         if (opt.isEmpty()) {
             throw new IllegalArgumentException("No vehicle with such ID");
         }
         this.vehicleRepository.deleteById(vehicleId);
-        return opt.get();
     }
 }
